+++ kotsu/error.py
@@ -1,0 +1,7 @@
+"""Exceptions and warnings."""
+
+
+class EntityIsDeprecated(ValueError):
+    """Raised when attempting to make an instance of a deprecated entity."""
+
+    pass