+++ kotsu/typing.py
@@ -20,9 +20,9 @@ Model = Any
 # A Validation is a callable (e.g. function) that takes a Model, and returns a Results.
 Validation = Callable[..., Results]
 
-# Validations can also take a Model _and_ an artefact directory, which is a string, which
-# Validations may use as the location to write any artefacts/data objects output by the Validation.
-# For storing Model artefacts (e.g. saving model state, or saving training history), Validations
-# should pass the `artefact_directory` to Models directly within the Validation, and Models should
-# be implemented such to use this arg as their output artefact store location.
-ValidationWithOutputArtefacts = Callable[[Model, str], Results]
+# Validations can also take a Model and two artefacts directory kwargs; `validation_artefacts_dir`
+# and`model_artefacts_dir`. Each can be used to write any artefacts/data objects formed by the
+# Validation or Model respectively. Passing `model_artefacts_dir` to Models is recommended for them
+# to use to store output artefacts (e.g. saving model state, or saving training history) to
+# directly.
+ValidationWithOutputArtefacts = Callable[[Model, str, str], Results]