+++ tests/test_end_to_end.py
@@ -1,17 +1,15 @@
 """An end to end example and test."""
 from typing import Callable
 
-import os
 import pickle
 
 from sklearn import datasets, svm
 from sklearn.model_selection import cross_validate
 
-from kotsu import run
-from kotsu.registration import ModelRegistry, ValidationRegistry
+import kotsu
 
 
-model_registry = ModelRegistry()
+model_registry = kotsu.registration.ModelRegistry()
 
 model_registry.register(
     id="SVC-v1",
@@ -25,23 +23,22 @@ model_registry.register(
     kwargs={"kernel": "linear", "C": 0.5, "random_state": 1},
 )
 
-validation_registry = ValidationRegistry()
+validation_registry = kotsu.registration.ValidationRegistry()
 
 
 def factory_iris_cross_val(folds: int) -> Callable:
     """Factory for iris cross validation."""
 
-    def iris_cross_val(model, artefacts_directory=None) -> dict:
+    def iris_cross_val(model, validation_artefacts_dir=None, model_artefacts_dir=None) -> dict:
         """Iris classification cross validation."""
         X, y = datasets.load_iris(return_X_y=True)
         scores = cross_validate(model, X, y, cv=folds, return_estimator=True)
 
-        if artefacts_directory:
-            os.makedirs(artefacts_directory)
+        if model_artefacts_dir:
             # Save the trained models from each fold
             for fold_idx, model in enumerate(scores["estimator"]):
-                with open(artefacts_directory + f"model_from_fold_{fold_idx}.pk", "wb") as handle:
-                    pickle.dump(model, handle)
+                with open(model_artefacts_dir + f"model_from_fold_{fold_idx}.pk", "wb") as f:
+                    pickle.dump(model, f)
 
         results = {f"fold_{i}_score": score for i, score in enumerate(scores["test_score"])}
         results["mean_score"] = scores["test_score"].mean()
@@ -65,10 +62,15 @@ validation_registry.register(
 
 
 def test_run(tmpdir):
-    run.run(model_registry, validation_registry, str(tmpdir) + "/validation_results.csv")
+    kotsu.run.run(
+        model_registry, validation_registry, results_path=str(tmpdir) + "/validation_results.csv"
+    )
 
 
 def test_run_save_models(tmpdir):
-    run.run(
-        model_registry, validation_registry, str(tmpdir) + "/validation_results.csv", str(tmpdir)
+    kotsu.run.run(
+        model_registry,
+        validation_registry,
+        results_path=str(tmpdir) + "/validation_results.csv",
+        artefacts_store_dir=str(tmpdir),
     )