+++ tests/test_run.py
@@ -18,6 +18,7 @@ class FakeRegistry:
         for id_ in ids:
             entity = mock.Mock()
             entity.id = id_
+            entity.deprecated = False
             entitys.append(entity)
             instance = mock.Mock()
             instance.return_value = {}
@@ -70,30 +71,34 @@ def test_form_results(mocker, tmpdir):
     assert patched_store_write.call_args[0][1] == results_path
 
 
-@pytest.mark.parametrize("artefacts_store_directory", [None, "test_dir/"])
-def test_validation_calls(artefacts_store_directory, mocker):
+@pytest.mark.parametrize("artefacts_store_dir", [None, "test_dir/"])
+def test_validation_calls(artefacts_store_dir, mocker):
     _ = mocker.patch("kotsu.store.write")
 
-    models = ["model_1", "model_2"]
+    models = ["model_1", "model_2", "model_3"]
     model_registry = FakeRegistry(models)
-    validations = ["validation_1"]
+    model_registry.entitys[1].deprecated = True
+    validations = ["validation_1", "validation_2"]
     validation_registry = FakeRegistry(validations)
+    validation_registry.entitys[1].deprecated = True
 
     kotsu.run.run(
         model_registry,
         validation_registry,
-        artefacts_store_directory=artefacts_store_directory,
+        artefacts_store_dir=artefacts_store_dir,
     )
-    if artefacts_store_directory is not None:
+    if artefacts_store_dir is not None:
         validation_registry.instances[0].assert_has_calls(
             [
                 mock.call(
                     model_registry.instances[0],
-                    artefacts_directory=f"{artefacts_store_directory}validation_1/model_1/",
+                    validation_artefacts_dir=f"{artefacts_store_dir}validation_1/",
+                    model_artefacts_dir=f"{artefacts_store_dir}validation_1/model_1/",
                 ),
                 mock.call(
-                    model_registry.instances[1],
-                    artefacts_directory=f"{artefacts_store_directory}validation_1/model_2/",
+                    model_registry.instances[2],
+                    validation_artefacts_dir=f"{artefacts_store_dir}validation_1/",
+                    model_artefacts_dir=f"{artefacts_store_dir}validation_1/model_3/",
                 ),
             ]
         )
@@ -104,24 +109,44 @@ def test_validation_calls(artefacts_store_directory, mocker):
                     model_registry.instances[0],
                 ),
                 mock.call(
-                    model_registry.instances[1],
+                    model_registry.instances[2],
                 ),
             ]
         )
+    validation_registry.instances[1].assert_not_called()
 
 
-@pytest.mark.parametrize("skip_if_prior_result", [True, False])
-def test_skip_if_prior_result(skip_if_prior_result, mocker, tmpdir):
+def test_passing_run_params_to_validation(mocker):
+    _ = mocker.patch("kotsu.store.write")
+
+    models = ["model_1"]
+    model_registry = FakeRegistry(models)
+    validations = ["validation_1"]
+    validation_registry = FakeRegistry(validations)
+
+    kotsu.run.run(
+        model_registry,
+        validation_registry,
+        run_params={"test_param": "test_param_value"},
+    )
+    validation_registry.instances[0].assert_has_calls(
+        [mock.call(model_registry.instances[0], test_param="test_param_value")]
+    )
+
+
+@pytest.mark.parametrize("force_rerun", [None, ["model_1"], "all"])
+def test_force_rerun(force_rerun, mocker, tmpdir):
     patched_run_validation_model = mocker.patch(
         "kotsu.run._run_validation_model",
         side_effect=[
-            ({"test_result": "result"}, 10),
+            ({"test_result": "result_1"}, 10),
             ({"test_result": "result_2"}, 20),
             ({"test_result": "result_3"}, 30),
+            ({"test_result": "result_4"}, 40),
         ],
     )
 
-    models = ["model_1"]
+    models = ["model_1", "model_2"]
     model_registry = FakeRegistry(models)
     validations = ["validation_1"]
     validation_registry = FakeRegistry(validations)
@@ -131,8 +156,9 @@ def test_skip_if_prior_result(skip_if_prior_result, mocker, tmpdir):
         model_registry,
         validation_registry,
         results_path=results_path,
-        skip_if_prior_result=skip_if_prior_result,
+        force_rerun=force_rerun,
     )
+    out_df = pd.read_csv(results_path)
 
     results_df = pd.DataFrame(
         [
@@ -140,7 +166,7 @@ def test_skip_if_prior_result(skip_if_prior_result, mocker, tmpdir):
                 "validation_id": "validation_1",
                 "model_id": "model_1",
                 "runtime_secs": 10,
-                "test_result": "result",
+                "test_result": "result_1",
             },
             {
                 "validation_id": "validation_1",
@@ -151,44 +177,59 @@ def test_skip_if_prior_result(skip_if_prior_result, mocker, tmpdir):
         ]
     )
 
-    out_df = pd.read_csv(results_path)
-
-    assert patched_run_validation_model.call_count == 1
-    pd.testing.assert_frame_equal(out_df, results_df.iloc[:1])
-
-    models = ["model_1", "model_2"]
-    model_registry = FakeRegistry(models)
+    assert patched_run_validation_model.call_count == 2
+    pd.testing.assert_frame_equal(out_df, results_df)
 
     kotsu.run.run(
         model_registry,
         validation_registry,
         results_path=results_path,
-        skip_if_prior_result=skip_if_prior_result,
+        force_rerun=force_rerun,
     )
-
     out_df = pd.read_csv(results_path)
 
-    if skip_if_prior_result:
+    if force_rerun is None:
         assert patched_run_validation_model.call_count == 2
         pd.testing.assert_frame_equal(out_df, results_df)
-    elif not skip_if_prior_result:
+
+    if isinstance(force_rerun, list) and force_rerun[0] == "model_1":
         results_df = pd.DataFrame(
             [
                 {
                     "validation_id": "validation_1",
                     "model_id": "model_1",
-                    "runtime_secs": 20,
-                    "test_result": "result_2",
+                    "runtime_secs": 30,
+                    "test_result": "result_3",
                 },
                 {
                     "validation_id": "validation_1",
                     "model_id": "model_2",
-                    "runtime_secs": 30,
-                    "test_result": "result_3",
+                    "runtime_secs": 20,
+                    "test_result": "result_2",
                 },
             ]
         )
         assert patched_run_validation_model.call_count == 3
+        pd.testing.assert_frame_equal(out_df, results_df)
+
+    if force_rerun == "all":
+        results_df = pd.DataFrame(
+            [
+                {
+                    "validation_id": "validation_1",
+                    "model_id": "model_1",
+                    "runtime_secs": 30,
+                    "test_result": "result_3",
+                },
+                {
+                    "validation_id": "validation_1",
+                    "model_id": "model_2",
+                    "runtime_secs": 40,
+                    "test_result": "result_4",
+                },
+            ]
+        )
+        assert patched_run_validation_model.call_count == 4
         pd.testing.assert_frame_equal(out_df, results_df)